premierleaguewebproject/requirements.txt
Cheetah==2.4.4
Django==1.9.5
Landscape-Client==14.12
MySQL-python==1.2.5
PAM==0.4.2
Pillow==3.2.0
PyYAML==3.10
SecretStorage==2.0.0
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
amqp==1.4.9
anyjson==0.3.3
apt-xapian-index==0.45
argparse==1.2.1
billiard==3.3.0.23
bzr==2.7.0dev1
celery==3.1.18
chardet==2.0.1
cloud-init==0.7.5
colorama==0.2.5
configobj==4.7.2
django-background-task==0.1.8
django-background-tasks==1.0.11
django-compat==1.0.10
django-request==1.4.0
html5lib==0.999
httplib2==0.8
importlib==1.0.3
jsonpatch==1.3
jsonpointer==1.0
keyring==3.5
kombu==3.0.35
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
oauth==1.0.1
paramiko==1.10.1
pbr==1.9.1
prettytable==0.7.2
pyOpenSSL==0.13
pycrypto==2.6.1
pycurl==7.19.3
pygobject==3.12.0
pyserial==2.6
python-apt==0.9.3.5ubuntu2
python-debian==0.1.21-nmu2ubuntu2
python-geoip==1.2
python-geoip-geolite2==2015.0303
pytz==2016.3
requests==2.2.1
simplejson==3.3.1
six==1.10.0
ssh-import-id==3.21
stevedore==1.12.0
urllib3==1.7.1
virtualenv==15.0.1
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
wadllib==1.3.2
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.0.5
